qr_code/requirements.txt
qrcode[pil]
python-dotenv